requirements.txt
discord.py==2.4.0
boto3==1.35.5
openai==1.42.0